src/requirements.txt
pytest
numpy
Pillow
plotly
